requirements.txt
requests>=2.28.1
darwin-py>=0.8.57
numpy>=1.21.2
jsonschema>=4.2.1
Pillow>=8.4.0
argparse>=1.4.0
flask>=2.0.2
boto3>=1.20.23
SimpleITK>=2.1.1
fire>=0.4.0
nibabel>=3.2.1
matplotlib>=3.5.1